added HL7 Monitor and ONVIF CAMERA page
